Create Working on Matrices
Task 1. Working with Matrices
A square matrix 𝑀 is called orthogonal if 𝑀𝑇𝑀 = 𝐼. Common examples of orthogonal matrices are matrices that represent rotations and reflections.
(1)	Give an nontrivial example of an orthogonal matrix 𝑀. Write numpy code to check that the columns of 𝑀 (when interpreted as vectors) are unit vectors (magnitude of 1) and every pair of columns is orthogonal (perpendicular). Also illustrate (using numpy and matplotlib) that when 𝑀 is used as a matrix transformation, it is an isometry, i.e., it preserves both magnitudes of vectors and angles between vectors. 
	I used ortho_group to generate an orthogonal matrix (M MT = I), and splitted the matrix columns in to vectors and calculated magnitude of each column which is 1 (unit vectors).
	Also checked the orthogonal condition for each column .
	Implemented matrix transformation for the orthogonal matrix which represents rotations and reflections.